Moves Blokk to Cell.Konnectable
