--- react/components/Blokk/components/Embed/index.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-import Generic from 'react/components/Blokk/components/Generic';
-import Connectable from 'react/components/Blokk/components/Connectable';
-
-export default class Embed extends Connectable {
-  render() {
-    const { embed, ...rest } = this.props;
-
-    return (
-      <Generic src={embed.src} alt={embed.title} {...rest} />
-    );
-  }
-}
