--- react/components/Blokk/components/Metadata/fragments/modelMetadata.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  fragment ModelMetadata on Model {
-    updated_at(relative: true)
-  }
-`;
