+++ react/components/Cell/components/Konnectable/components/Image/index.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+import { propType } from 'graphql-anywhere';
+
+import imageFragment from 'react/components/Cell/components/Konnectable/components/Image/fragments/image';
+
+import Generic from 'react/components/Cell/components/Konnectable/components/Generic';
+
+export default class Image extends Component {
+  static propTypes = {
+    image: propType(imageFragment).isRequired,
+  }
+
+
+  render() {
+    const { image, ...rest } = this.props;
+
+    return (
+      <Generic src={image.src} alt={image.title} {...rest} />
+    );
+  }
+}
