+++ react/components/Cell/components/Konnectable/components/Img/index.js
