+++ react/components/Cell/components/Konnectable/components/Metadata/index.js
@@ -1,0 +1,80 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { propType } from 'graphql-anywhere';
+
+import metadataFragment from 'react/components/Cell/components/Konnectable/components/Metadata/fragments/metadata';
+
+import Box from 'react/components/UI/Box';
+import Text from 'react/components/UI/Text';
+import Truncate from 'react/components/UI/Truncate';
+import MetadataFileExtension from 'react/components/Cell/components/Konnectable/components/Metadata/components/MetadataFileExtension';
+import MetadataPlay from 'react/components/Cell/components/Konnectable/components/Metadata/components/MetadataPlay';
+
+export default class Metadata extends Component {
+  static propTypes = {
+    konnectable: propType(metadataFragment).isRequired,
+    mode: PropTypes.oneOf(['resting', 'hover', 'overlay']),
+  }
+
+  static defaultProps = {
+    mode: 'resting',
+  }
+
+  render() {
+    const {
+      mode,
+      konnectable: {
+        __typename,
+        title,
+        user,
+        connection,
+        updated_at,
+        file_extension,
+      },
+    } = this.props;
+
+    return (
+      <Box my={5} px={5}>
+        <Text f={1} fontWeight="bold" textAlign="center" color="gray.medium">
+          {mode === 'resting' &&
+            <div>
+              <Truncate length={40}>
+                {title}
+              </Truncate>
+
+              {__typename === 'Attachment' &&
+                <MetadataFileExtension ext={file_extension} />
+              }
+
+              {__typename === 'Embed' &&
+                <MetadataPlay />
+              }
+            </div>
+          }
+
+          {mode === 'hover' &&
+            <div>
+              {connection &&
+                <div>
+                  {user.id === connection.user.id
+                    ? <div>Added by {user.name}</div>
+                    : <div>Connected by {connection.user.name}</div>
+                  }
+                  {connection.created_at}
+                </div>
+              }
+
+              {!connection &&
+                <div>
+                  Added by {user.name}
+                  <br />
+                  {updated_at}
+                </div>
+              }
+            </div>
+          }
+        </Text>
+      </Box>
+    );
+  }
+}
