+++ react/components/Cell/components/Konnectable/queries/blokk.js
@@ -1,0 +1,12 @@
+import gql from 'graphql-tag';
+
+import konnectableFragment from 'react/components/Cell/components/Konnectable/fragments/konnectable';
+
+export default gql`
+  query Blokk($id: ID!) {
+    blokk(id: $id) {
+      ... Konnectable
+    }
+  }
+  ${konnectableFragment}
+`;
