+++ react/components/Cell/index.js
@@ -1,5 +1,7 @@
+import Konnectable from 'react/components/Cell/components/Konnectable';
 import Identifiable from 'react/components/Cell/components/Identifiable';
 
 export default {
+  Konnectable,
   Identifiable,
 };
