+++ react/components/ProfileContents/fragments/profileContents.js
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag';
 
-import blokkFragment from 'react/components/Blokk/fragments/blokk';
+import konnectableCellFragment from 'react/components/Cell/components/Konnectable/fragments/konnectableCell';
 
 export default gql`
   fragment ProfileContents on Identifiable {
@@ -8,9 +8,9 @@ export default gql`
       __typename
       id: slug
       contents: kontents(type: $type, per: $per, page: $page, direction: DESC, sort_by: $sort) {
-        ... Blokk
+        ... KonnectableCell
       }
     }
   }
-  ${blokkFragment}
+  ${konnectableCellFragment}
 `;
